refactor(api): centralize application wiring
